fix(detector): 修复后处理中的浮点数精度问题和安全检查
- 将所有浮点数字面量从 double 转换为 float（添加 f 后缀）
- 在输出长度计算中使用 size_t 类型避免整数溢出
- 修复 anchor 类型从 uint32 到 float32 的转换
- 修正 max_det 边界条件判断（> 改为 >=）
- 添加输出和结果为空的安全检查
- 在 ObjectBuffer 中修复 off-by-one 断言错误
- 添加关键点解析时的越界保护

+++ lib/detector/NetOutput.hpp
@@ -231,7 +231,7 @@ class NetOutput
      * @description: 获取缓冲区大小, 方便调试
      * @return
      */
-    uint32 get_buffer_size() const { return this->buffer.size(); }
+    size_t get_buffer_size() const { return this->buffer.size(); }
 
     /***
      * @description: 打印数据信息
@@ -241,8 +241,8 @@ class NetOutput
     {
         // 将各种信息打印为字符串
         std::string info = format_string(
-            "batch_size: %d, channel: %d, height: %d, width: %d, buffer_size: %d, "
-            "buffer_capacity: %d, buffer_data*: %p",
+            "batch_size: %d, channel: %d, height: %d, width: %d, buffer_size: %zu, "
+            "buffer_capacity: %zu, buffer_data*: %p",
             this->batch_size, this->channel, this->height, this->width, this->buffer.size(), this->buffer.capacity(),
             this->buffer.data());
